Moved pseudoRandom32CharHex.js into utils folder
